// src/interfaces/InformationPage.tsx
export interface InformationPage {
  id?: number;
  title: string;
  title_2: string;

  subtitle_1: string;
  description_1: string;

  subtitle_2: string;
  description_2: string;

  subtitle_3: string;
  description_3: string;

  image_1: string;

  subtitle_4: string;
  description_4: string;

  image_2: string;

  subtitle_5: string;
  description_5: string;

  subtitle_6: string;
  description_6: string;

  subtitle_7: string;
  description_7: string;

  image_3: string;

  created_at?: string;
  updated_at?: string;
}